Refactor Events and Scope
