+++ internal/event/Events.go
@@ -1,4 +1,4 @@
-package internal
+package event
 
 import "fmt"
